+++ packages/analytics-js-common/src/constants/integrations/Braze/constants.ts
@@ -1,6 +1,6 @@
+import { BRAZE_NAME as NAME, BRAZE_DISPLAY_NAME as DISPLAY_NAME } from '../Destinations';
+
 const DIR_NAME = 'Braze';
-const NAME = 'BRAZE';
-const DISPLAY_NAME = 'Braze';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {