+++ packages/analytics-js-common/src/constants/integrations/Bugsnag/constants.ts
@@ -1,6 +1,6 @@
+import { BUGSNAG_NAME as NAME, BUGSNAG_DISPLAY_NAME as DISPLAY_NAME } from '../Destinations';
+
 const DIR_NAME = 'Bugsnag';
-const NAME = 'BUGSNAG';
-const DISPLAY_NAME = 'Bugsnag';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {