+++ packages/analytics-js-common/src/constants/integrations/HubSpot/constants.ts
@@ -1,6 +1,6 @@
+import { HS_NAME as NAME, HS_DISPLAY_NAME as DISPLAY_NAME } from '../Destinations';
+
 const DIR_NAME = 'HubSpot';
-const NAME = 'HS';
-const DISPLAY_NAME = 'HubSpot';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {