+++ packages/analytics-js-common/src/constants/integrations/Mixpanel/constants.ts
@@ -1,6 +1,6 @@
+import { MP_NAME as NAME, MP_DISPLAY_NAME as DISPLAY_NAME } from '../Destinations';
+
 const DIR_NAME = 'Mixpanel';
-const NAME = 'MP';
-const DISPLAY_NAME = 'Mixpanel';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {