+++ packages/analytics-js-common/src/constants/integrations/Optimizely/constants.ts
@@ -1,6 +1,6 @@
+import { OPTIMIZELY_NAME as NAME, OPTIMIZELY_DISPLAY_NAME as DISPLAY_NAME } from '../Destinations';
+
 const DIR_NAME = 'Optimizely';
-const NAME = 'OPTIMIZELY';
-const DISPLAY_NAME = 'Optimizely Web';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {