+++ packages/analytics-js-common/src/constants/integrations/SpotifyPixel/constants.ts
@@ -1,6 +1,9 @@
+import {
+  SPOTIFYPIXEL_NAME as NAME,
+  SPOTIFYPIXEL_DISPLAY_NAME as DISPLAY_NAME,
+} from '../Destinations';
+
 const DIR_NAME = 'SpotifyPixel';
-const NAME = 'SPOTIFYPIXEL';
-const DISPLAY_NAME = 'Spotify Pixel';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {